+++ programs/lb_clmm/src/lib.rs
@@ -55,7 +55,10 @@ use instructions::withdraw_protocol_fee::*;
 #[cfg(feature = "localnet")]
 declare_id!("LbVRzDTvBDEcrthxfZ4RL6yiq3uZw8bS6MwtdY6UhFQ");
 
-#[cfg(not(feature = "localnet"))]
+#[cfg(feature = "staging")]
+declare_id!("tLBro6JJuZNnpoad3p8pXKohE9f7f7tBZJpaeh6pXt1");
+
+#[cfg(not(any(feature = "localnet", feature = "staging")))]
 declare_id!("LBUZKhRxPF3XUpBCjp4YzTKgLccjZhTSDM9YuVaPwxo");
 
 pub mod admin {